pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sshq"
version = "0.1.0"
description = "AI-assisted SSH wrapper"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "flask",
    "google-genai"
]

[project.scripts]
sshq = "sshq.cli:main"
